BUG#123456789: Fixed CMake and consequent build warnings:
  - Fixed erroneous set of CMAKE_CXX_FLAGS;
  - Added -Wno-unused-private-field GCC flag;
  - Fixed -Wshadow, -Wconversion and -Wextra compilation warnings;

+++ src/fabric_cache/src/fabric_cache_plugin.cc
@@ -70,7 +70,7 @@ const string prompt_password(const string &prompt) {
   std::cout << prompt << ": ";
 
   // prevent showing input
-  console.c_lflag &= ~ECHO;
+  console.c_lflag &= ~(uint)ECHO;
   tcsetattr(STDIN_FILENO, TCSANOW, &console);
 
   string result;
